backend/requirements.txt
flask
flask-cors
requests
python-dotenv
transformers
torch  
pillow  
accelerate
bitsandbytes